Project Update
- Updated module to work as a factory worker to other factories and
functions.

- Added the following facilities (factories):

Core
	The core serves as the factory worker for all of its tools and
facilities.  This includes .type(), .extend(), .each(), among all the
others listed below. 
	
	The module may be fed an application configuration by instantiating a
new instance ( i.e. new nodakwaeri( config ); ).  .init() may then be
invoked with or without any additional parameters (none for a default
nodakwaeri application).  Users can opt to provide an alternative
router, controller, renderer and/or rendering tools, and/or
database/model system.  The alternative configurations are not available
at this time as they are still under development.
	
Server
	The server creates the HTTP object and handles dispatch to the router
for further processing.  Developers will eventually be able to configure
the router or provide an alternative router.
	
Router
	The router determines whether a request is for media or for an
application resource or page, and handles the request by either feeding
the media or by handling dispatch to the controller for further
processing based on the referrer and the type as mentioned.
	
Controller
	The controller acts as the 'base' controller class and provides
mechanisms for developers to minimize the amount of code needed in the
application's controllers.

Renderer
	The renderer is the layout and view system.  It allows developers an
easy way to design their application's pages and content, maintaining a
high level of reusability and organization by allowing them to separate
layout and content.
	  
	The renderer also provides a preprocessor for a baked in scripting
language termed 'kml' or 'kwaeri markup language'.  The preprocessor
interprets subscripts within kml files which make up the html content,
and is given access to any supplied models so they can be seamlessly
integrated into the output by supplying the model within the klay object
in the controller.  I opted for kml instead of khtml for various
reasons.

Html (Rendering Tools)
	The HTML rendering tools provide the output for all the kml html tools.
This includes everything from html.label() for generating labels to
html.submit() for creating a submit button.  This is the only facility
where 'this' didn't work inside so the code is slightly different from
the rest of them.  I have been rushing so am probably overlooking
something but at least I'm reminding myself here.
	
	The HTML rendering tools are passed the klay object and can make use of
the model member to incorporate the model/data system.
	
	The rendering tools are separated this way so that in the future third
party developers can create tools that could replace or extend the baked
in rendering tools allowing them to truly customize the nodakwaeri
application framework to their tastes and needs.

- The model system is coming soon, I've built a MySQL connector (c++)
add-on (obviously c++), and have it working and tested on both 64-bit
linux and windows.  The issue is simply that I needed to wait until I
had more of the framework complete before I could design the remaining
facilities that would provide seamless integration with the framework.
I'm actively working on the model system at this point and it will be
finished within the next couple of days.

- Everything is working and can be tested as is with the current MMod
Node.js Application Framework Example http://github.com/mmod/nk-mvc

- Updated version to 0.0.2 (still experimental at this point)

- Creating new tag and publishing to NPM as nk or nodakwaeri (whichever
is available first)

Change-Id: I2b41322ad79fcc845f33360c96b424f6a363ebbf
Signed-off-by: Richard Winters <rik@mmogp.com>