Project Update
- Current state, includes:

Session

The session facility manages the client sessions, allowing connections
to act persistently so that users do not need to continuously log in
over and over.

- The framework now properly handles processing of GET and POST data,
with check in place so that the server is not flooded and/or not too
easily suseptable to DOS attacks

- The layout system now properly renders model data (at least labels and
display names, nodamysql is currently in progress and once it is
complete enough the model system will be added as well as the 'for'
script flow for templating database records and other chunks of data.


Change-Id: I79328cd16c91411a482ad27e2c47a51a8b57d449
Signed-off-by: Richard Winters <rik@mmogp.com>