Project Update
- Fixed the session facilities.  It is now secure, makes use of both a
session identifier as well as a session differentiator, to ensure
identity of a client in the even that one is able to forge a session
identifier which exists server-side.  It also now properly creates
cookies and supports all options for cookies (expiration, path, host,
secure)
- Added session tools.  Developers can now make use of res.session.set
and req.session.get.
- Added res.redirect, so users can redirect.
- Rewrote the hash method to properly create a seeded/salted hash from a
string (256-bit SHA)
- Remove crypt directory, going to let SJCL roam as its own while
keeping it bootstrapped right in.  Moved bsd license and sjcl library to
sjcl directory.

Change-Id: I38d1d1841b1b38ac334268a0c1432085463e910c
Signed-off-by: Richard Winters <rik@mmogp.com>