Project Update
- .each() now works on objects or arrays, and will pass the key and
value to the callback, just the value, or even nothing but execute the
callback, depending on the number of arguments specified.

- .extend() now works on objects and arrays, no need to worry about
passing only one or the other around - but the same rules apply all
around regarding how properties are extended in objects and arrays.  

If extending an object with an array:
If the numerical properties exist they are skipped, if they do not exist
they are added on.  

If extending an array with an object
By default the object's properties are added to new indexes of the
array. Passing false as a third parameter is optional and allows one to
specify to nk that it should not new indexes to an array being extended
by an object but rather set any non existing numeric indexes that exist
in the object as properties.

- Changed tabs policy to 'Spaces Only', since tabs really really are
evil.

- Updated README

- Updated version

Change-Id: I7cacc9d611a54f1d092024482a813b4dc7d800ad
Signed-off-by: Rikkotec <rik@mmogp.com>